Refactor pool deletion hook into useDeleteZpool
