+++ src/hooks/useDeleteZpool.ts
@@ -1,5 +1,6 @@
+import { useCallback, useState } from 'react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import type { ZpoolQueryResult } from '../@types/zpool';
+import type { ZpoolCapacityEntry, ZpoolQueryResult } from '../@types/zpool';
 import axiosInstance from '../lib/axiosInstance';
 
 interface DeleteZpoolPayload {
@@ -16,18 +17,27 @@ const deleteZpool = async ({ name }: DeleteZpoolPayload) => {
   const response = await axiosInstance.delete<DeleteZpoolResponse>(
     '/api/zpool/delete',
     {
-      params: { name },
-      data: { name },
+      data: { pool_name: name },
     }
   );
 
   return response.data;
 };
 
-export const useDeleteZpool = () => {
-  const queryClient = useQueryClient();
+interface UseDeleteZpoolOptions {
+  onSuccess?: (poolName: string) => void;
+  onError?: (error: Error, poolName: string) => void;
+}
 
-  return useMutation<DeleteZpoolResponse, Error, DeleteZpoolPayload>({
+export const useDeleteZpool = ({
+  onSuccess,
+  onError,
+}: UseDeleteZpoolOptions = {}) => {
+  const queryClient = useQueryClient();
+  const [targetPool, setTargetPool] = useState<ZpoolCapacityEntry | null>(null);
+  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+
+  const deleteMutation = useMutation<DeleteZpoolResponse, Error, DeleteZpoolPayload>({
     mutationFn: deleteZpool,
     onSuccess: (_data, variables) => {
       queryClient.setQueryData<ZpoolQueryResult | undefined>(
@@ -50,4 +60,48 @@ export const useDeleteZpool = () => {
       queryClient.invalidateQueries({ queryKey: ['zpool'] });
     },
   });
+
+  const requestDelete = useCallback((pool: ZpoolCapacityEntry) => {
+    setErrorMessage(null);
+    setTargetPool(pool);
+  }, []);
+
+  const handleClose = useCallback(() => {
+    setTargetPool(null);
+    setErrorMessage(null);
+  }, []);
+
+  const confirmDelete = useCallback(() => {
+    if (!targetPool || deleteMutation.isPending) {
+      return;
+    }
+
+    setErrorMessage(null);
+
+    deleteMutation.mutate(
+      { name: targetPool.name },
+      {
+        onSuccess: () => {
+          onSuccess?.(targetPool.name);
+          handleClose();
+        },
+        onError: (error) => {
+          setErrorMessage(error.message);
+          onError?.(error, targetPool.name);
+        },
+      }
+    );
+  }, [deleteMutation, handleClose, onError, onSuccess, targetPool]);
+
+  return {
+    isOpen: Boolean(targetPool),
+    targetPool,
+    requestDelete,
+    closeModal: handleClose,
+    confirmDelete,
+    isDeleting: deleteMutation.isPending,
+    errorMessage,
+  };
 };
+
+export type UseDeleteZpoolReturn = ReturnType<typeof useDeleteZpool>;
